Update library to allow for same version for training and evaluation

+++ rtdetr_pytorch/src/nn/backbone/common.py
@@ -20,7 +20,7 @@ class ConvNormLayer(nn.Module):
         self.act = nn.Identity() if act is None else get_activation(act) 
 
     def forward(self, x):
-        return self.act(self.norm(self.conv(x)))
+        return self.act(self.norm(self.conv(x.float())))
 
 
 class FrozenBatchNorm2d(nn.Module):
